// src/shared/models/shared-user.model.ts
import { UserStatus } from 'src/shared/constants/auth.constant'
import { z } from 'zod'

export const userSchema = z.object({
    id: z.number(),
    email: z.string().email(),
    name: z.string().min(1).max(100),
    password: z.string().min(6).max(100),
    phoneNumber:  z.string().min(9).max(15),
    avatar: z.string().nullable(),
    status: z.enum([UserStatus.ACTIVE, UserStatus.INACTIVE, UserStatus.BLOCKED]),
    roleId: z.number().positive(),
    createdById: z.number().nullable(),
    updatedById: z.number().nullable(),
    deletedAt: z.date().nullable(),
    createdAt: z.date(),
    updatedAt: z.date(),
    totpSecret: z.boolean()

})
export type UserType = z.infer<typeof userSchema>